Rearchitectures database interface classes
Directly accessing the classes themselves should be unnecessary now;
the BaseDB class now contains a factory to automatically generate
methods for interacting with data from tables that a given subclass has
access to. This factory is wrapped by functions to generate initialized
database interfaces, available in the global scope in the store module.