Implements wrapper classes for ta-lib functions
For each function in ta_func.h, there is a wrapper class, which holds
the arguments that the library function takes as private members, has a
create function to generate new objects, and a run method to run the
appropriate ta-lib function with the object's private data members.
This may be used to implement reflection for ta-lib by creating a
(static) mapping between function names (strings) and function pointers
to the wrapping class' create method.