Implements bot architecture
Rewrites the basic architecture of the bots; old bots.* are now stored
as bots.*.old, for reference, as not all features from the old code have
equivalents in the rewrite. All code now compiles, although it is still
untested. Reflection is fully implemented, but will not be needed until
we have to write actual rules. The model for the bot table was made more
detailed; a host (ip address) field was added.