Adds global-level library directory
I'm going to start writing the server code, and while it seems
appropriate to create a directory at the global project level for it,
the server code will need access to the client libraries, so I made a
global-level library directory, and the Makefile now symlinks each local
lib directory into the global one (e.g., bots/lib -> lib/bots). The
Makefile assumes it's running in a submodule directory, and that the
"global" level is one level up. I also added __init__.py's so that the
Cython extensions meant for use with Django are esaily accessible