Implements non-blocking server architecture
Uses TNonblockingServer (throttles processor threads using a pool) and
TCompactProtocol (faster protocol) to implement the client/server test
programs. Now that I have the client/server code up and running, I'm
going to start implementing the actual architecture; the server should
be basically the same, although I'll probably wrap it into a class for
ease of access, and the client is going to become a library used for
connecting to the server. control.* (included here as some skeleton
code) will define a controller that automatically handles load balancing
using Chandy-Misra, and the Cython interface will be changed to use the
controller as a backend.