Fixes compilation on droplet
The -march=native flag was causing the build process to fail with the
error:

CPU you selected does not support x86-64 instruction set

on the droplet. This seems like it could be an issue with the native CPU
architecture (i.e., that of the actual server); for now, I've removed
the flag.