Adds server-side bot code
I refactored some of the framework code to support the new division of
labor between the (Thrift) server and the the (bot) client. I also
replaced json serialization with archiving; rule vectors are now stored
as binary data in a NuoDB table. I made some changes to the Makefile to
fix the new build.