Changes params to CLOB in database
I changed params to a CLOB in cryptobot_rule and updated our libraries.
NuoDB's C++ API appears to be broken (unless there is a more fundamental
issue with our code) for the BINARY and BLOB types. CLOB also seems to
play nice with Django, so we can keep our SQL schemas centralized to
models.py. I changed params to a TextField with binary (base64)
encoding, which is implemented in NuoDB's Django backend as a CLOB.